added controller tests and associated changes in controllers to pass them

+++ app/controllers/comments_controller.rb
@@ -3,12 +3,14 @@ class CommentsController < ApplicationController
 
   def create
     if signed_in_user?
-      type = params[:comment][:commentable_type].classify
-      resource = type.constantize.find(params[:comment][:commentable_id])
-      if resource.comments.create(description: params[:comment][:description],user_id: current_user.id)
-        flash[:success] = "Comment has been added!"
+      if instance = comment_parent
+        if instance.save!
+          flash[:success] = "Comment has been added!"
+        else
+          flash[:error] = "Comment not added, returning to the nerdery to diagnose the problem. . . ."
+        end
       else
-        flash[:error] = "Comment not added, returning to the nerdery to diagnose the problem. . . ."
+        flash[:error] = "Could not find #{params[:comment][:commentable_type].downcase}"
       end
       redirect_to :back
     else
@@ -18,12 +20,16 @@ class CommentsController < ApplicationController
 
   def destroy
     if signed_in_user?
-      @comment = Comment.find(params[:id])
-      if @comment.destroy
-        flash[:success] = "Comment has been destroyed..."
-        redirect_to root_path
+      if @comment = Comment.find_by_id(params[:id])
+        if @comment.destroy
+          flash[:success] = "Comment has been destroyed..."
+          redirect_to root_path
+        else
+          flash[:error] = "Couldn't destroy the comment, it has gone sentient..."
+          redirect_to :back
+        end
       else
-        flash[:error] = "Couldn't destroy the comment, it has gone sentient..."
+        flash[:error] = "Couldn't find the comment. . ."
         redirect_to :back
       end
     else
@@ -35,6 +41,14 @@ class CommentsController < ApplicationController
 
     def white_listed_comment_params
       params.require(:comment).permit(:description)
+    end
+
+    def comment_parent
+      type = params[:comment][:commentable_type].classify
+      resource = type.constantize.
+                 find_by_id(params[:comment][:commentable_id])
+      return nil unless resource
+      resource.comments.new(description: params[:comment][:description], user_id: current_user.id)
     end
 
 end
